Polynomial multiplication with FFT.

+++ src/params.h
@@ -7,10 +7,10 @@ class Params
 {
 protected:
     Params() { }
-    Params(int B1_, int B2_) : B1(B1_), B2(B2_) { }
+    Params(uint64_t B1_, uint64_t B2_) : B1(B1_), B2(B2_) { }
     virtual ~Params() { }
 
-    double get_profit(int B1, int B2, int max_size, ProbSmooth& prob, double log_sieving_depth, double log_known_divisors, double primality_cost);
+    double get_profit(uint64_t B1, uint64_t B2, int max_size, ProbSmooth& prob, double log_sieving_depth, double log_known_divisors, double primality_cost);
     void find_pp1_optimal_bounds(int max_size, ProbSmooth& prob, double log_sieving_depth, double log_known_divisors, double primality_cost);
     void find_pp1_optimal_B2(int max_size, ProbSmooth& prob, double log_sieving_depth, double log_known_divisors, double primality_cost);
     void find_pp1_stage2_optimal_params(int max_size);
@@ -23,8 +23,8 @@ public:
     virtual int stage2_size();
 
 public:
-    int B1;
-    int B2;
+    uint64_t B1;
+    uint64_t B2;
     int D;
     int A;
     int L;
@@ -36,8 +36,8 @@ public:
 class PM1Params : public Params
 {
 public:
-    PM1Params(int B1, int B2, int max_size, bool poly);
-    PM1Params(int B1, int max_size, ProbSmooth& prob, double log_sieving_depth, double log_known_divisors, double primality_cost);
+    PM1Params(uint64_t B1, uint64_t B2, int max_size, bool poly);
+    PM1Params(uint64_t B1, int max_size, ProbSmooth& prob, double log_sieving_depth, double log_known_divisors, double primality_cost);
     PM1Params(int max_size, ProbSmooth& prob, double log_sieving_depth, double log_known_divisors, double primality_cost);
 
 };
@@ -45,8 +45,8 @@ public:
 class PP1Params : public Params
 {
 public:
-    PP1Params(int B1, int B2, int max_size, bool poly);
-    PP1Params(int B1, int max_size, ProbSmooth& prob, double log_sieving_depth, double log_known_divisors, double primality_cost);
+    PP1Params(uint64_t B1, uint64_t B2, int max_size, bool poly);
+    PP1Params(uint64_t B1, int max_size, ProbSmooth& prob, double log_sieving_depth, double log_known_divisors, double primality_cost);
     PP1Params(int max_size, ProbSmooth& prob, double log_sieving_depth, double log_known_divisors, double primality_cost);
 
     virtual int stage1_cost() override;
@@ -55,7 +55,7 @@ public:
 class EdECMParams : public Params
 {
 public:
-    EdECMParams(int B1, int B2, int max_size, bool poly);
+    EdECMParams(uint64_t B1, uint64_t B2, int max_size, bool poly);
 
     virtual int stage1_cost() override;
     virtual int stage1_size() override;
